+++ app/assets/stylesheets/horses_race.css
