+++ app/views/minigames/slotmachine_show.html.erb
@@ -1,3 +1,7 @@
+<% content_for :head do %>
+  <%= stylesheet_link_tag "slot_machine" %>
+<% end %>
+
 <h1>Slot Machine</h1>
 <div class="slot_machine">
   <div class="barra">
